LayoutTests/accessibility/img-aria-button-alt-tag.html

<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<script>
    if (window.layoutTestController)
        layoutTestController.dumpAsText();
</script>
<body id="body">
 
    <img id='image' src="resources/cake.png" role="button" alt="alternate" title="title">
    
    <p id="description"></p>
    <div id="console"></div>
     
    <script>
        description("This test makes sure that a img acting as an aria button still returns its alt tag as the description");
 
        if (window.accessibilityController) {
            var result = document.getElementById("result");

            var body = document.getElementById("body").focus();
            var imgUIElement = accessibilityController.focusedElement.childAtIndex(0).childAtIndex(0);
            shouldBe("imgUIElement.description", "'AXDescription: alternate'");
        }
        successfullyParsed = true;
    </script>
    
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
